Changed getPrevScope to getCurrentScope in the compoundMessagePronoun for PREV
compoundMessagePronoun only adds the scope to the scopestack AFTER checking what type of scope it refers to. In case of PREV being the second scope it can't find a previous scope since there will only be one scope in the stack. So it needs getCurrentScope to find the correct scope to refer to.
This doesn't matter for RHS mentions of PREV since those don't create a scope.